orders scripts in descending order and adds benchmark for selects with no index

+++ Projects/Index/B_Tree/Scripts/query_differences/query_differences.lua
@@ -1,4 +1,6 @@
 local con = sysbench.sql.driver():connect()
+local with_index = not (os.getenv("NO") == "index")
+
 function prepare()
     local create_kunden_query = [[
         CREATE TABLE KUNDEN (
@@ -20,13 +22,18 @@ function prepare()
     ]]
 
     con:query(create_kunden_query)
-    con:query(create_indices)
+
+    if with_index then
+        con:query(create_indices)
+    end
     print("Table 'KUNDEN' and index have been successfully created.")
 end
 
 
 function cleanup()
-    con:query("DROP INDEX combined_index ON KUNDEN;")
+    if with_index then
+        con:query("DROP INDEX combined_index ON KUNDEN;")
+    end
     con:query("DROP TABLE IF EXISTS KUNDEN;")
 
     print("Cleanup successfully done.")
